+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# dev
+
 # processx 3.2.1
 
 * processx does not depend on assertthat now, and the crayon package