+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# dev
+# processx 3.3.0
 
 * `process` can now redirect the standard error to the standard output, via
   specifying `stderr = "2>&1"`. This works both with files and pipes.